+++ cli/src/merge_tools/mod.rs
@@ -18,6 +18,9 @@ mod external;
 
 use std::sync::Arc;
 
+use bstr::BString;
+use itertools::Itertools;
+use jj_lib::backend::FileId;
 use jj_lib::backend::MergedTreeId;
 use jj_lib::config::ConfigGetError;
 use jj_lib::config::ConfigGetResultExt as _;
@@ -26,10 +29,13 @@ use jj_lib::conflicts::extract_as_single_hunk;
 use jj_lib::conflicts::ConflictMarkerStyle;
 use jj_lib::gitignore::GitIgnoreFile;
 use jj_lib::matchers::Matcher;
+use jj_lib::merge::Merge;
+use jj_lib::merge::MergedTreeValue;
 use jj_lib::merged_tree::MergedTree;
 use jj_lib::repo_path::InvalidRepoPathError;
 use jj_lib::repo_path::RepoPath;
 use jj_lib::repo_path::RepoPathBuf;
+use jj_lib::repo_path::RepoPathUiConverter;
 use jj_lib::settings::UserSettings;
 use jj_lib::working_copy::SnapshotError;
 use pollster::FutureExt;
@@ -97,8 +103,10 @@ pub enum ConflictResolveError {
          see the exact invocation)."
     )]
     EmptyOrUnchanged,
-    #[error("Backend error")]
+    #[error(transparent)]
     Backend(#[from] jj_lib::backend::BackendError),
+    #[error(transparent)]
+    Io(#[from] std::io::Error),
 }
 
 #[derive(Debug, Error)]
@@ -257,6 +265,14 @@ impl DiffEditor {
     }
 }
 
+/// A file to be merged by a merge tool.
+struct MergeToolFile {
+    repo_path: RepoPathBuf,
+    conflict: MergedTreeValue,
+    file_merge: Merge<Option<FileId>>,
+    content: Merge<BString>,
+}
+
 /// Configured 3-way merge editor.
 #[derive(Clone, Debug)]
 pub struct MergeEditor {
@@ -312,41 +328,58 @@ impl MergeEditor {
     /// Starts a merge editor for the specified file.
     pub fn edit_file(
         &self,
+        ui: &Ui,
+        path_converter: &RepoPathUiConverter,
         tree: &MergedTree,
-        repo_path: &RepoPath,
+        repo_paths: &[&RepoPath],
     ) -> Result<MergedTreeId, ConflictResolveError> {
-        let conflict = match tree.path_value(repo_path)?.into_resolved() {
-            Err(conflict) => conflict,
-            Ok(Some(_)) => return Err(ConflictResolveError::NotAConflict(repo_path.to_owned())),
-            Ok(None) => return Err(ConflictResolveError::PathNotFound(repo_path.to_owned())),
-        };
-        let file_merge = conflict.to_file_merge().ok_or_else(|| {
-            let summary = conflict.describe();
-            ConflictResolveError::NotNormalFiles(repo_path.to_owned(), summary)
-        })?;
-        let simplified_file_merge = file_merge.clone().simplify();
-        // We only support conflicts with 2 sides (3-way conflicts)
-        if simplified_file_merge.num_sides() > 2 {
-            return Err(ConflictResolveError::ConflictTooComplicated {
-                path: repo_path.to_owned(),
-                sides: simplified_file_merge.num_sides(),
-            });
-        };
-        let content =
-            extract_as_single_hunk(&simplified_file_merge, tree.store(), repo_path).block_on()?;
+        let merge_tool_files: Vec<MergeToolFile> = repo_paths
+            .iter()
+            .map(|&repo_path| {
+                let conflict = match tree.path_value(repo_path)?.into_resolved() {
+                    Err(conflict) => conflict,
+                    Ok(Some(_)) => {
+                        return Err(ConflictResolveError::NotAConflict(repo_path.to_owned()))
+                    }
+                    Ok(None) => {
+                        return Err(ConflictResolveError::PathNotFound(repo_path.to_owned()))
+                    }
+                };
+                let file_merge = conflict.to_file_merge().ok_or_else(|| {
+                    let summary = conflict.describe();
+                    ConflictResolveError::NotNormalFiles(repo_path.to_owned(), summary)
+                })?;
+                let simplified_file_merge = file_merge.clone().simplify();
+                // We only support conflicts with 2 sides (3-way conflicts)
+                if simplified_file_merge.num_sides() > 2 {
+                    return Err(ConflictResolveError::ConflictTooComplicated {
+                        path: repo_path.to_owned(),
+                        sides: simplified_file_merge.num_sides(),
+                    });
+                };
+                let content =
+                    extract_as_single_hunk(&simplified_file_merge, tree.store(), repo_path)
+                        .block_on()?;
+                Ok(MergeToolFile {
+                    repo_path: repo_path.to_owned(),
+                    conflict,
+                    file_merge,
+                    content,
+                })
+            })
+            .try_collect()?;
 
         match &self.tool {
             MergeTool::Builtin => {
-                let tree_id = edit_merge_builtin(tree, repo_path, content).map_err(Box::new)?;
+                let tree_id = edit_merge_builtin(tree, &merge_tool_files).map_err(Box::new)?;
                 Ok(tree_id)
             }
             MergeTool::External(editor) => external::run_mergetool_external(
+                ui,
+                path_converter,
                 editor,
-                file_merge,
-                content,
-                repo_path,
-                conflict,
                 tree,
+                &merge_tool_files,
                 self.conflict_marker_style,
             ),
         }